Implement a simple asynchronous serializer
AsyncSerializer writes to ByteBuffer so long there is enough bytes to
write the next value. The remaining write operations are saved in a
Runnable callback, and can be repeated once the buffer is freed.

AsyncSerializer reads from ByteBuffer so long the buffer contains enough
bytes to read the next value. The remaining read operations are,
similarly, saved in a Runnable callback.

This prevents buffer overflow and underflow even for very small buffer
sizes (as small as 4 bytes).

A roundtrip serialization test is added.